refactor(pdf): isolate caption figure envelopes
